+++ core/pipeline-driver-queue/lib/consts/component-name.js
@@ -6,6 +6,7 @@ module.exports = {
     MAIN: 'MAIN',
     ETCD_PERSISTENT: 'ETCD_PERSISTENT',
     JOBS_PRODUCER: 'JOBS_PRODUCER',
+    CONCURRENCY: 'CONCURRENCY',
     JOBS_CONSUMER: 'JOBS_CONSUMER',
     AGGREGATION_METRIC: 'AGGREGATION_METRIC',
     CONSUMER: 'CONSUMER',
